+++ README.md
@@ -67,6 +67,48 @@ Lombok: 1.18.30
 
 
 
+#### POST /users/passreset
+
+​	Request
+
+```json
+{
+    "cemail":"email address",
+    "cpassword":"password",
+    "securityQuestion":"question",
+    "answer":"answer",
+}
+```
+
+​	Response
+
+```
+S:
+{
+    "code": 200,
+    "msg": "Password reset success",
+    "data": 1
+}
+F:
+{
+    "code": 400,
+    "msg": "No Account Found",
+    "data": 0
+}
+{
+    "code": 400,
+    "msg": "Wrong security question or answer",
+    "data": 0
+}
+{
+    "code": 400,
+    "msg": "Something went wrong when trying to reset password",
+    "data": 0
+}
+```
+
+
+
 #### POST /transactions/transfer
 
 ​	Request
@@ -88,6 +130,43 @@ Lombok: 1.18.30
   
 }
 ```
+
+#### POST /transactions/withdraw
+
+Checking and saving all use this one.
+
+​	Request
+
+```json
+{
+    "anum":1,
+    "abalance":100.00
+}
+```
+
+​	Response
+
+```
+success:
+{
+	"code": 200,
+    "msg": "withdraw success",
+    "data": 100.00 						//the new balance
+}
+Fail:
+{
+	"code": 400,
+    "msg": "Something went wrong when withdrawing",
+    "data": 0
+}
+{
+	"code": 400,
+    "msg": "Balance not enough",
+    "data": 0
+}
+```
+
+
