+++ safe-server/src/main/java/team/ybj/controller/UserController.java
@@ -2,11 +2,10 @@ package team.ybj.controller;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
-import team.ybj.exception.UsernameDuplicatedException;
-import team.ybj.pojo.ResponseResult;
-import team.ybj.pojo.YbjCustomer;
-import team.ybj.service.LoginService;
-import team.ybj.service.RegService;
+import team.ybj.exception.*;
+import team.ybj.pojo.*;
+import team.ybj.service.*;
+
 
 
 @RestController
@@ -18,6 +17,10 @@ public class UserController {
 
     @Autowired
     RegService regService;
+
+    @Autowired
+    PasswordService PasswordService;
+
 
     @PostMapping("login")
     @ResponseBody
@@ -36,6 +39,22 @@ public class UserController {
             responseResult = new ResponseResult(400, "Duplicate username", 1);
         }catch (RuntimeException e) {
             responseResult = new ResponseResult(401, "Register failed for some unknown reason", 1);
+        }
+        return responseResult;
+    }
+
+        @PostMapping("passreset")
+    @ResponseBody
+    public ResponseResult PassReset(@RequestBody YbjCustomer customer) {
+        ResponseResult responseResult;
+        try {
+            responseResult = PasswordService.PassReset(customer);
+        }catch (UsernameDuplicatedException e){
+            responseResult = new ResponseResult(400, "No Account Found", 0);
+        }catch (ServiceException e) {
+            responseResult = new ResponseResult(400, "Wrong security question or answer", 0);
+        }catch (PassResetException e) {
+            responseResult = new ResponseResult(400, "Something went wrong when trying to reset password", 0);
         }
         return responseResult;
     }
